requirements.txt
django==3.1
djangorestframework==3.11.1
django-auth-adfs==1.3.1
psycopg2-binary==2.8.5
gunicorn==20.0.4
python-dotenv==0.14.0
reportlab==3.5.47